Refactor stuff
